+++ src/blocks/place-card/place-card.css
