+++ src/index.html
@@ -5,12 +5,12 @@
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <meta http-equiv="X-UA-Compatible" content="ie=edge">
   <title>Mesto</title>
-  <link rel="stylesheet" href="./pages/index.css">
+  <link rel="stylesheet" href="<%=htmlWebpackPlugin.files.chunks.main.css %>">
 </head>
 <body>
   <div class="root">
     <header class="header root__section">
-      <img src="./images/logo.svg" alt="mesto logo" class="logo">
+      <img src="<%=require('./images/logo.svg')%>" alt="mesto logo" class="logo">
     </header>
     <div class="profile root__section">
       <div class="user-info">
@@ -25,24 +25,10 @@
     </div>
     <div class="places-list root__section">
 
-      <!-- Пример вёрстки карточки: 
-
-      <div class="place-card">
-        <div class="place-card__image" style="background-image: url(https://pictures.s3.yandex.net/frontend-developer/cards-compressed/kamchatka.jpg)">
-          <button class="place-card__delete-icon"></button>
-        </div>
-        <div class="place-card__description">
-          <h3 class="place-card__name">Камчатка</h3>
-          <button class="place-card__like-icon"></button>
-        </div>
-      </div>
-
-       конец примера. -->
-
     </div>
     <div class="popup">
       <div class="popup__content">
-        <img src="./images/close.svg" alt="" class="popup__close">
+        <img src="<%=require('./images/close.svg')%>" alt="" class="popup__close">
         <h3 class="popup__title">Новое место</h3>
         <form id="form" novalidate class="popup__form" name="new">
             <input 
@@ -71,7 +57,7 @@
 
     <div class="popup popup__edit-profile">
       <div class="popup__content">
-        <img src="./images/close.svg" alt="" class="popup__edit-profile__close">
+        <img src="<%=require('./images/close.svg')%>" alt="" class="popup__edit-profile__close">
         <h3 class="popup__title">Редактировать профиль</h3>
         <form id="form" novalidate class="popup__form" name="edit">
             <input 
@@ -101,12 +87,12 @@
 
     <div class="popup popup__image">
       <div class="popup__content popup__content_image">
-        <img src="./images/close.svg" alt="" class="popup__image_close">
+        <img src="<%=require('./images/close.svg')%>" alt="" class="popup__image_close">
         <img alt="" class="popup__image_fullsize">
       </div>
     </div>
 
   </div>
-  <script src="script.js"></script>
+  <script src="<%= htmlWebpackPlugin.files.chunks.main.entry %>"></script>
 </body>
 </html>
