+++ src/index.js
@@ -1,0 +1,6 @@
+import "./style.css";
+
+import './script';
+
+
+const serverUrl = NODE_ENV === 'development' ? 'http://praktikum.tk/cohort4' : 'https://praktikum.tk/cohort4';
